finished jummah notifications
