+++ code/backend/app.js
@@ -16,7 +16,7 @@ const routes = require($DIR + '/routing/index.js')
 const dbSettings =  require($DIR + '/database/settings.js')
 
 const createRootUserScript = require('./cron/createRootUser.js')
-const wednesdayNotifications = require('./cron/wednesdayNotification.js')
+const jummahNotifications = require('./cron/jummahNotifications.js')
 
 const PORT = 80
 const DATABASE = process.env.DATABASE || 'mongodb://localhost:27017/khateebRemind'
@@ -53,6 +53,7 @@ db.once('open', () => { console.log(`${dbType} Mongo is listening`) })
 db.on('error', (error) => { console.log(`Connection error : ${error}`) })
 app.listen(PORT, () => { console.log(`App is listening on port ${PORT}`) })
 
-if (process.env.NODE_ENV === 'production')
+if (process.env.NODE_ENV === 'production') {
     createRootUserScript(() => { console.log(`Root User Creation Script Initiated`) })
-wednesdayNotifications(() => { console.log(`Set jummah notifications!`) })
+    jummahNotifications(() => { console.log(`Set jummah notifications!`) })
+}
